requirements.txt
asgiref==3.3.0
Django==3.0
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
importlib-metadata==2.0.0
Markdown==3.3.3
pymongo==3.11.0
pytz==2020.4
sqlparse==0.2.4
zipp==3.4.0
